Fixing errors - delete movies

+++ db/migrate/20220921173240_create_reviews.rb
@@ -2,7 +2,7 @@ class CreateReviews < ActiveRecord::Migration[7.0]
   def change
     create_table :reviews do |t|
       t.text :comment
-      t.decimal :rating
+      t.integer :rating
       t.references :list, null: false, foreign_key: true
 
       t.timestamps
